+++ api/cache.go
@@ -1,4 +1,4 @@
-package data
+package api
 
 import (
 	"fmt"
