+++ cmd/cacheclean.go
@@ -1,7 +1,7 @@
 package cmd
 
 import (
-	"github.com/IvanKuzyshyn/aoc-go/data"
+	"github.com/IvanKuzyshyn/aoc-go/api"
 	"github.com/spf13/cobra"
 )
 
@@ -24,7 +24,7 @@ func NewCacheCleanCommand() *cobra.Command {
 }
 
 func (c *cacheCleanCommand) runE(command *cobra.Command, args []string) error {
-	cache := data.Cache{
+	cache := api.Cache{
 		Dir: "cache",
 	}
